Fixed data type under builtins (int->uint)
https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I7LLDJ

+++ ecmascript/dfx/hprof/heap_profiler.cpp
@@ -54,7 +54,7 @@ bool HeapProfiler::DumpHeapSnapshot(DumpFormat dumpFormat, Stream *stream, Progr
     LOG_ECMA(INFO) << "HeapProfiler DumpSnapshot start";
     size_t heapSize = vm_->GetHeap()->GetHeapObjectSize();
     LOG_ECMA(ERROR) << "HeapProfiler DumpSnapshot heap size " << heapSize;
-    int32_t heapCount = vm_->GetHeap()->GetHeapObjectCount();
+    int32_t heapCount = static_cast<int32_t>(vm_->GetHeap()->GetHeapObjectCount());
     if (progress != nullptr) {
         progress->ReportProgress(0, heapCount);
     }
@@ -111,7 +111,7 @@ bool HeapProfiler::StopHeapTracking(Stream *stream, Progress *progress, bool new
     if (heapTracker_ == nullptr) {
         return false;
     }
-    int32_t heapCount = vm_->GetHeap()->GetHeapObjectCount();
+    int32_t heapCount = static_cast<int32_t>(vm_->GetHeap()->GetHeapObjectCount());
 
     const_cast<EcmaVM *>(vm_)->StopHeapTracking();
     if (newThread) {
